requirements.txt
fastapi==0.115.8
SQLAlchemy==2.0.38
alembic==1.14.1
asyncpg==0.30.0
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
pyjwt==2.10.1
httpx==0.28.1
python-multipart==0.0.20
pytest-asyncio==0.25.3
pytest-mock==3.14.0
faker==37.0.0
factory-boy==3.3.3
respx==0.22.0